🧹 Cleanup code and add console messages

+++ src/main/services/preferences.ts
@@ -1,22 +1,28 @@
 import { existsSync, writeFileSync } from 'fs';
-import { app } from 'electron';
 import { resolve } from 'path';
 
 import { DEFAULT_PREFERENCES_OBJECT } from '../../shared/models/default-preferences';
+import { startStorageService } from './storage';
 
-let userData = app.getPath('userData')
+import colors from 'colors';
+
+let userData = startStorageService()
 
 export const preferencesExist = () => { 
     return existsSync(resolve(userData, 'preferences.json'))
 }
 
-export const preferencesFirstSetup = () => {
-    const defaultPreferences = DEFAULT_PREFERENCES_OBJECT
+export const preferencesLoad = () => {
+    console.log(`${colors.blue.bold('[Preferences]')} Loaded preferences file located at ${userData}`);
 
-    writeFileSync(
-        resolve(userData, 'preferences.json'),
-        JSON.stringify(defaultPreferences)
-    )
+    if(preferencesExist() == false) {
+        const defaultPreferences = DEFAULT_PREFERENCES_OBJECT
+
+        writeFileSync(
+            resolve(userData, 'preferences.json'),
+            JSON.stringify(defaultPreferences)
+        )
+    }
 }
 
 export const preferencesLocation = resolve(userData, 'preferences.json');
